Refactored non Sanic code into core
