+++ core/mongo/document.py
