+++ core/search/filter_functions.py
@@ -1,5 +1,4 @@
-from typing import List
-from server.search.type_filter import *
+from core.search.type_filter import *
 
 
 def type_filter_function(type_filter: TypeFilter) -> List[dict]:
