+++ core/search/sort_by.py
@@ -1,4 +1,4 @@
-from server.search import fields
+from core.search import fields
 from enum import Enum
 
 from typing import List
