+++ core/users/session.py
@@ -4,9 +4,9 @@ from bson import ObjectId
 from sanic.request import Request
 
 from server.app import BaseModel
-from server.mongo.document import Document
+from core.mongo.document import Document
 
-from server.word_embedding.sanic_supervised_models import load_model, SupervisedModels
+from core.word_embedding.sanic_supervised_models import load_model, SupervisedModels
 
 
 class Session(BaseModel, Document):
