+++ core/word_embedding/utils.py
